Flytter oppdatering av utsatt til dato endepunkt til oppslag og oppdatering controller
